First

+++ src/main/java/com/controller/AmenityController.java
@@ -50,7 +50,7 @@ public class AmenityController {
 
 
     @GetMapping("/{amenityId}")
-    public ResponseEntity<Object> getAmenityById(@PathVariable("amenityId") Long amenityId) {
+    public ResponseEntity<Object> getAmenityById(@PathVariable("amenityId") int amenityId) {
         Amenity amenity = amenityService.getAmenityById(amenityId);
         if (amenity == null) {
             return ResponseEntity.status(404).body("{\"code\": \"GETFAILS\", \"message\": \"Amenity doesn't exist\"}");
@@ -59,7 +59,7 @@ public class AmenityController {
     }
 
     @PutMapping("/update/{AmenityId}")
-    public ResponseEntity<Object> updateAmenity(@PathVariable("AmenityId") Long amenityId, @RequestBody Amenity amenity) {
+    public ResponseEntity<Object> updateAmenity(@PathVariable("AmenityId") int amenityId, @RequestBody Amenity amenity) {
         if (!amenityService.findById(amenityId)) {
             return ResponseEntity.status(404).body("{\"code\": \"UPDTFAILS\", \"message\": \"Amenity doesn't exist\"}");
         }
@@ -70,7 +70,7 @@ public class AmenityController {
    
     
     @DeleteMapping("/{amenityId}")
-    public ResponseEntity<Object> deleteAmenity(@PathVariable("amenityId") Long id) {
+    public ResponseEntity<Object> deleteAmenity(@PathVariable("amenityId") int id) {
         if (!amenityService.existsById(id)) {
             return ResponseEntity.status(404).body("{\"code\": \"DLTFAILS\", \"message\": \"Amenity doesn't exist exist\"}");
         }
@@ -84,7 +84,7 @@ public class AmenityController {
 
 
     @GetMapping("/hotel/{hotel_id}")
-    public ResponseEntity<?> getAmenitiesByHotel(@PathVariable Long hotel_id) {
+    public ResponseEntity<?> getAmenitiesByHotel(@PathVariable int hotel_id) {
         List<Amenity> amenities = hotelamenityService.getAmenitiesByHotel(hotel_id);
         if (amenities.isEmpty()) {
             return ResponseEntity.status(404).body("{\"code\": \"GETALLFAILS\", \"message\": \"hotel not found with given hotel id}");
@@ -93,7 +93,7 @@ public class AmenityController {
     }
 
     @GetMapping("/room/{room_id}")
-    public ResponseEntity<?> getAmenitiesByRoom(@PathVariable Long room_id) {
+    public ResponseEntity<?> getAmenitiesByRoom(@PathVariable int room_id) {
         List<Amenity> amenities = roomamenityService.getAmenitiesByRoom(room_id);
         if (amenities.isEmpty()) {
             return ResponseEntity.status(404).body("{\"code\": \"GETALLFAILS\", \"message\": \"room not found with given hotel id}");
