+++ src/main/java/com/dao/HotelDAO.java
@@ -9,8 +9,9 @@ import com.model.Hotel;
 import com.model.Room;
 
 @Repository
-public interface HotelDAO extends JpaRepository<Hotel, Long> {
+public interface HotelDAO extends JpaRepository<Hotel, Integer> {
 	
     @Query("SELECT r FROM Hotel r WHERE r.location = :location")
     List<Room> findHotelsByLocation(String location);
+    
 }
