+++ src/main/java/com/dao/PaymentDAO.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.model.Payment;
 @Repository
-public interface PaymentDAO extends JpaRepository<Payment, Long> {
+public interface PaymentDAO extends JpaRepository<Payment, Integer> {
 
 	   List<Payment> findByPaymentStatus(String status);
 }
