+++ src/main/java/com/dao/ReviewDAO.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import com.model.Review;
 @Repository
-public interface ReviewDAO extends JpaRepository<Review, Long> {
+public interface ReviewDAO extends JpaRepository<Review, Integer> {
 	
 	
     List<Review> findByRating(int rating);
